removed or combined redundant redux code
