--- client/src/redux/actions/behavior.js
@@ -1,7 +1,0 @@
-import { OPEN_BEHAVIOR } from '../reducers/behavior';
-
-export const openBehavior = () => ({
-  type: OPEN_BEHAVIOR,
-});
-
-export default openBehavior;
