--- client/src/redux/actions/navigation.js
@@ -1,6 +1,0 @@
-export const SET_NAVBAR_SECTION = 'SET_NAVBAR_SECTION';
-
-export const setNavbarSection = section => ({
-  type: SET_NAVBAR_SECTION,
-  payload: section,
-});
