+++ client/src/redux/actions/onboarding.js
@@ -5,6 +5,34 @@ export const SET_DEMOGRAPHICS_COMORBIDITIES = 'SET_DEMOGRAPHICS_COMORBIDITIES';
 
 export const RESET_DEMOGRAPHICS_COMORBIDITIES = 'RESET_DEMOGRAPHICS_COMORBIDITIES';
 
+export const DISCLAIMER_ANSWER = 'DISCLAIMER_ANSWER';
+
+export const RESET_ANSWER = 'RESET_ANSWER';
+
+// action creators
+export function setDisclaimerAnswer(answer) {
+  return {
+    type: DISCLAIMER_ANSWER,
+    answer,
+  };
+}
+
+export function resetDisclaimerAnswer() {
+  return {
+    type: RESET_ANSWER,
+  };
+}
+
+// thunk
+export const setDisclaimerAnswerThunk = (answer, disclaimerAnswerJSON, userSession) => async dispatch => {
+  await userSession
+    .putFile(`disclaimer.json`, JSON.stringify(disclaimerAnswerJSON))
+    .then(() => 200)
+    .catch(err => console.error(err));
+
+  dispatch(setDisclaimerAnswer(answer));
+};
+
 export function setDemographicsComorbidities(formData) {
   return {
     type: SET_DEMOGRAPHICS_COMORBIDITIES,
