+++ client/src/tests/Chart.test.jsx
@@ -7,13 +7,6 @@ import Chart from '../components/Chart';
 
 const mockStore = configureMockStore([thunk]);
 
-const mockChartJsState = {
-  temperature: {
-    labels: ['2020-02-28', '2020-03-2', '2020-03-4', '2020-03-5', '2020-03-9'],
-    values: [80, 70, 56, 90, 55],
-  },
-};
-
 const mockObservationsReducerState = {
   observations: [],
 };
@@ -21,7 +14,6 @@ const mockObservationsReducerState = {
 describe('Chart', () => {
   it('should render a Chart component ', () => {
     const store = mockStore({
-      temperatureReducer: mockChartJsState,
       observationsReducer: mockObservationsReducerState,
     });
     window.HTMLCanvasElement.prototype.getContext = () => {};
